fix width for location icon

+++ app/src/main/java/n7/ad2/utils/extension/Int.kt
@@ -4,8 +4,8 @@ package n7.ad2.utils.extension
 
 import android.content.res.Resources
 
-val Int.toPx: Float
-    get() = (this * Resources.getSystem().displayMetrics.density)
+val Int.toPx: Int
+    get() = (this * Resources.getSystem().displayMetrics.density.toInt())
 
-val Int.toDp: Float
-    get() = (this / Resources.getSystem().displayMetrics.density)
+val Int.toDp: Int
+    get() = (this / Resources.getSystem().displayMetrics.density.toInt())
